Moved arearangespline series to subfolder
