+++ js/masters/highcharts-more.src.js
@@ -11,7 +11,7 @@
 import '../Extensions/Pane.js';
 import '../Core/Axis/RadialAxis.js';
 import '../Series/AreaRange/AreaRangeSeries.js';
-import '../Series/AreaSplineRangeSeries.js';
+import '../Series/AreaSplineRange/AreaSplineRangeSeries.js';
 import '../Series/ColumnRangeSeries.js';
 import '../Series/ColumnPyramid/ColumnPyramidSeries.js';
 import '../Series/GaugeSeries.js';
